+++ docs/changelog.md
@@ -627,9 +627,11 @@ Many feature gates are removed in order to make most features make always availa
 - new macro: `fmtcat`.
 - new trait: `DebugExt`.
 - new types: `FmtNum`, `FmtNumConf`, `FmtNumShape`, `FmtNumSign`, `FmtWriter`.
+- re-export `FromFn` as `FmtFromFn`.
 - move to [base]:
   - macros: `format_buf!`.
 - remove vendored `numtoa` crate, `NumToStr` trait replaced with `Digits` struct.
+- add method `Fmt::from_fn`.
 
 ### grapheme
 - new types: `GraphemeKind`, `GraphemeScanner`.